+++ CHANGELOG.md
@@ -1,11 +1,15 @@
+2.010
+ * Extensible let - hlet and helpers make using the primitive overloads of clojure functions easier.
+   See the ham-fisted.hlet and ham-fisted.primtive-invoke interfaces.
+
 2.009
  * typed 'nth' methods efficient for primitive manipulations - 'dnth', 'fnth', 'inth', 'lnth'.
- 
- 
+
+
 # 2.008
  * Custom reduce implemented for object array wrappers.
 
- 
+
 2.007
 * reduce namespace now has helper to create a parallel reducer.
 * hashset has optimized addall pathway when input is another hashset.